Look + item command, look inventory command.
now player can look at items and the program will print their
description (only if the item is in the room or in the player's
inventory). Player can also check what his inventory contains and also
whenever a description of a room is printed by the program it also
prints the list of items that are in that room.